fix bug data

+++ app/src/main/java/com/toshiro/weatherapp/data/network/historicalWeather/DataHistorical.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class DataHistorical(
     @SerializedName("clouds")
-    val clouds: Int,
+    val clouds: Double,
     @SerializedName("datetime")
     val datetime: String,
     @SerializedName("dewpt")
@@ -23,33 +23,33 @@ data class DataHistorical(
     @SerializedName("max_ghi")
     val maxGhi: Double,
     @SerializedName("max_temp")
-    val maxTemp: Int,
+    val maxTemp: Double,
     @SerializedName("max_temp_ts")
-    val maxTempTs: Int,
+    val maxTempTs: Double,
     @SerializedName("max_uv")
     val maxUv: Double,
     @SerializedName("max_wind_dir")
-    val maxWindDir: Int,
+    val maxWindDir: Double,
     @SerializedName("max_wind_spd")
     val maxWindSpd: Double,
     @SerializedName("max_wind_spd_ts")
-    val maxWindSpdTs: Int,
+    val maxWindSpdTs: Double,
     @SerializedName("min_temp")
-    val minTemp: Int,
+    val minTemp: Double,
     @SerializedName("min_temp_ts")
-    val minTempTs: Int,
+    val minTempTs: Double,
     @SerializedName("precip")
     val precip: Double,
     @SerializedName("precip_gpm")
-    val precipGpm: Int,
+    val precipGpm: Double,
     @SerializedName("pres")
-    val pres: Int,
+    val pres: Double,
     @SerializedName("rh")
     val rh: Double,
     @SerializedName("slp")
-    val slp: Int,
+    val slp: Double,
     @SerializedName("snow")
-    val snow: Int,
+    val snow: Double,
     @SerializedName("snow_depth")
     val snowDepth: Any,
     @SerializedName("solar_rad")
@@ -65,9 +65,9 @@ data class DataHistorical(
     @SerializedName("temp")
     val temp: Double,
     @SerializedName("ts")
-    val ts: Int,
+    val ts: Double,
     @SerializedName("wind_dir")
-    val windDir: Int,
+    val windDir: Double,
     @SerializedName("wind_gust_spd")
     val windGustSpd: Double,
     @SerializedName("wind_spd")
